refactor: 코드 리팩토링
AuthenticationPrincipal 사용해서 userId 가져오는지 테스트 하는 코드 작성
패키지 구성 변경
