+++ .gitignore
@@ -37,3 +37,5 @@ out/
 .vscode/
 
 src/main/resources/.env
+src/main/resources/application.properties
+src/main/resources/application-local.properties
