+++ src/main/java/org/example/postory/domain/auth/security/SecurityConfig.java
@@ -1,8 +1,8 @@
 package org.example.postory.domain.auth.security;
 
 import lombok.RequiredArgsConstructor;
-import org.example.postory.domain.auth.JwtProvider;
-import org.example.postory.domain.jwt.JwtAuthenticationFilter;
+import org.example.postory.domain.auth.jwt.JwtAuthenticationFilter;
+import org.example.postory.domain.auth.jwt.JwtProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
