+++ src/main/java/org/example/postory/domain/auth/service/AuthService.java
@@ -1,13 +1,14 @@
 package org.example.postory.domain.auth.service;
 
 import lombok.RequiredArgsConstructor;
-import org.example.postory.domain.auth.JwtProvider;
+import org.example.postory.domain.auth.jwt.JwtProvider;
 import org.example.postory.domain.auth.dto.AuthRequestDto;
-import org.example.postory.domain.auth.dto.JwtToken;
+import org.example.postory.domain.auth.jwt.JwtToken;
 import org.example.postory.domain.user.entity.User;
 import org.example.postory.domain.user.service.UserService;
 import org.example.postory.global.error.ApiException;
 import org.example.postory.global.error.response.ErrorType;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -33,7 +34,8 @@ public class AuthService {
     /**
      * accessToken을 재발급
      */
-    public JwtToken reIssue(String refreshToken){
+    public JwtToken reIssue(String bearerToken){
+        String refreshToken = bearerToken.replace("Bearer ", "");
         return jwtProvider.refreshToken(refreshToken);
     }
 }
