+++ src/main/java/org/example/postory/domain/user/controller/UserController.java
@@ -2,6 +2,8 @@ package org.example.postory.domain.user.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -11,7 +13,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
     @GetMapping
-    public ResponseEntity<String> test(){
-        return ResponseEntity.status(HttpStatus.OK).body("Test success");
+    public ResponseEntity<String> test(@AuthenticationPrincipal UserDetails userDetails){
+        return ResponseEntity.status(HttpStatus.OK).body(userDetails.getUsername());
     }
 }
