+++ src/main/resources/application-dev.properties
@@ -1,0 +1,11 @@
+spring.application.name=postory
+
+spring.datasource.url=jdbc:mysql://localhost:3306/postory
+spring.datasource.username=root
+spring.datasource.password=1212
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
+
+jwt.secret=qwrmlasfmlqw2491u4291urjeiqowfjkwklfnlksdnfi13fnou3nounf13
